parser: refactoring parser to pkg
Moved the parser package to pkg, as it cleans up/removes invalid import
cycles.
